+++ files/gitlab-ctl-commands-ee/lib/pgbouncer.rb
@@ -1,3 +1,4 @@
+require 'chef/mixins'
 require 'erb'
 
 # For testing purposes, if the first path cannot be found load the second
@@ -9,7 +10,7 @@ end
 
 module Pgbouncer
   class Databases
-    attr_accessor :install_path, :databases, :options, :ini_file, :json_file, :template_file, :attributes, :userinfo
+    attr_accessor :install_path, :databases, :options, :ini_file, :json_file, :template_file, :attributes, :userinfo, :database
     attr_reader :data_path
 
     def initialize(options, install_path, base_data_path)
@@ -20,12 +21,22 @@ module Pgbouncer
       @ini_file = options['databases_ini'] || attributes['gitlab']['pgbouncer']['databases_ini']
       @json_file = options['databases_json'] || attributes['gitlab']['pgbouncer']['databases_json']
       @template_file = "#{@install_path}/embedded/cookbooks/gitlab-ee/templates/default/databases.ini.erb"
+      @database = if attributes.key?('gitlab')
+                    attributes['gitlab']['gitlab-rails']['db_database']
+                  else
+                    'gitlabhq_production'
+                  end
       @databases = update_databases(JSON.parse(File.read(@json_file)))
       @userinfo = GitlabCtl::Util.userinfo(options['host_user']) if options['host_user']
     end
 
     def update_databases(original)
-      updated = {}
+      if original.empty?
+        original = {
+          database => {}
+        }
+      end
+      updated = Chef::Mixin::DeepMerge.merge(updated, original)
       original.each do |db, settings|
         updated[db] = ''
         settings['host'] = options['newhost'] if options['newhost']