+++ CHANGELOG.md
@@ -25,10 +25,11 @@ omnibus-gitlab repository.
 9.5.5
 
 - Add more options to repmgr.conf template
-- Update pgbouncer to use one style of logging 
+- Update pgbouncer to use one style of logging
 - Set bootstrap_expect to default to 3 for consul servers
 - Restart repmgrd after nodes are registered with a cluster
 - Add --node option to gitlab-ctl repmgr standby unregister
+- Fix bug where pgb-notify would fail when databases.json was empty
 
 9.5.4